Update installation to manual
The old avogadro installation from the ubuntu packages was not the
best version, so now the installation is made from source.
The Dockerfile was splitted into two parts, the base and avogadro
(copied this from sharelatex), so the librearies are installed in a
separate container.
The eigen3.2 version is installed, which is the last one that supports
eigen2, needed in avogadro.
A compose dockerfile was added to launch the container with
`docker compose up` in the future.